+++ test/GCClassic/integration/integrationTestCreate.sh
@@ -51,7 +51,7 @@ quick="${3}"
 thisDir=$(pwd -P)
 cd "${thisDir}"
 
-# GCClassic superproject directory
+# GCClassic superproject directory (absolute paths)
 cd ../../../../../
 superProjectDir=$(pwd -P)
 cd ${superProjectDir}
@@ -75,28 +75,66 @@ head_hco=$(export GIT_DISCOVERY_ACROSS_FILESYSTEM=1; \
 printf "${SEP_MAJOR}\n"
 printf "Creating GEOS-Chem Classic Integration Tests\n\n"
 printf "GCClassic #${head_gcc}\n"
-if [[ ! -d "${itRoot}/build" ]]; then
-    for dir in ${EXE_GCC_BUILD_LIST[@]}; do
-	echo " ... ${itRoot}/build/${dir}"
-	mkdir -p "${itRoot}/build/${dir}"
-    done
-fi
+printf "GEOS-Chem #${head_gc}\n"
+printf "HEMCO     #${head_hco}\n"
+printf "${SEP_MAJOR}\n"
 
-# Copying the run scripts to the integration test root folder
-printf "\nCopying run scripts to: ${itRoot}\n"
-cp -f ${envFile}                            ${itRoot}/gcclassic.env
-cp -f ${thisDir}/integrationTest*.sh        ${itRoot}
-cp -f ${thisDir}/commonFunctionsForTests.sh ${itRoot}
+#=============================================================================
+# Create integration test folder and subdirectories
+#=============================================================================
+
+# Create local convenience variables
+binDir="${itRoot}/${BIN_DIR}"
+buildDir="${itRoot}/${BUILD_DIR}"
+envDir="${itRoot}/${ENV_DIR}"
+execDir="${itRoot}/${EXEC_DIR}"
+logsDir="${itRoot}/${LOGS_DIR}"
+scriptsDir="${itRoot}/${SCRIPTS_DIR}"
+rundirsDir="${itRoot}/${RUNDIRS_DIR}"
+
+# Remove run directories in the test folder
+cleanup_files "${itRoot}"
+
+# Subdir for CMake builds (note: will create ${itRoot}
+printf "\nCreating CMake build directories:\n"
+for dir in ${EXE_GCC_BUILD_LIST[@]}; do
+    printf " ... ${buildDir}/${dir}\n"
+    mkdir -p "${buildDir}/${dir}"
+done
+
+# Subdir for executables
+printf "\nCreating exe files directory ${binDir}\n"
+mkdir -p "${binDir}"
+
+# Subdir for env files
+printf "Creating env files directory ${envDir}\n"
+mkdir -p "${envDir}"
+
+# Subdir for log files
+printf "Creating logs directory      ${logsDir}\n"
+mkdir -p "${logsDir}"
+
+# Subdir for scripts
+printf "Creating scripts directory   ${scriptsDir}\n"
+mkdir -p "${scriptsDir}"
+
+# Subdir for run directories
+printf "Creating rundirs directory   ${rundirsDir}\n"
+mkdir -p "${rundirsDir}"
+
+#=============================================================================
+# Copy files to the proper folders
+#=============================================================================
+
+printf "\nCopying run scripts to: ${itRoot}/${SCRIPTS_DIR}\n"
+cp -f ${envFile}                            ${itRoot}/${ENV_DIR}/gcclassic.env
+cp -f ${thisDir}/integrationTest*.sh        ${itRoot}/${SCRIPTS_DIR}
+cp -f ${thisDir}/commonFunctionsForTests.sh ${itRoot}/${SCRIPTS_DIR}
+cp -f ${thisDir)/README.template.md         ${itRoot}/README.md
 
 # Create a symbolic link to the code from the Integration Test itRoot folder
 [[ -L ${itRoot}/CodeDir ]] && unlink ${itRoot}/CodeDir
 ln -s "${superProjectDir}" ${itRoot}/CodeDir
-
-# Create log directory
-if [[ !(-d "${itRoot}/logs") ]]; then
-    printf "\nCreating log directory: ${itRoot}/logs\n"
-    mkdir "${itRoot}/logs"
-fi
 
 # Log file with echoback from rundir creation
 log="${itRoot}/logs/createIntegrationTests.log"
@@ -112,21 +150,21 @@ printf "\nCreating new run directories:\n"
 #=============================================================================
 
 # 4x5 merra2 CH4
-create_rundir "3\n1\n1\n1\n${itRoot}\n\nn\n" "${log}" "${itRoot}"
+create_rundir "3\n1\n1\n1\n${rundirsDir}\n\nn\n" "${log}"
 
 # 4x5 merra2 CO2
-create_rundir "4\n1\n1\n1\n${itRoot}\n\nn\n" "${log}" "${itRoot}"
+create_rundir "4\n1\n1\n1\n${rundirsDir}\n\nn\n" "${log}"
 
 # 4x5 merra2 aerosol
-create_rundir "2\n1\n1\n1\n${itRoot}\n\nn\n" "${log}" "${itRoot}"
+create_rundir "2\n1\n1\n1\n${rundirsDir}\n\nn\n" "${log}"
 
 # 4x5 merra2 carbon
-create_rundir "12\n1\n1\n1\n${itRoot}\n\nn\n" "${log}" "${itRoot}"
+create_rundir "12\n1\n1\n1\n${rundirsDir}\n\nn\n" "${log}"
 
 # 4x5 merra2 fullchem
-create_rundir "1\n1\n1\n1\n1\n${itRoot}\n\nn\n" "${log}" "${itRoot}"
+create_rundir "1\n1\n1\n1\n1\n${rundirsDir}\n\nn\n" "${log}"
 
-# DEBUG: Exit after creating a couple of rundirs if $quick is "yes"
+# DEBUG: Exit after creating a couple of rundirsDirs if $quick is "yes"
 if [[ "x${quick}" == "xyes" ]]; then
     cd ${thisDir}
     exit 0
@@ -134,76 +172,76 @@ fi
 
 # 4x5 merra2 fullchem_LuoWd
 dir="gc_4x5_merra2_fullchem_LuoWd"
-create_rundir "1\n1\n1\n1\n1\n${itRoot}\n${dir}\nn\n" "${log}" "${itRoot}"
+create_rundir "1\n1\n1\n1\n1\n${rundirsDir}\n${dir}\nn\n" "${log}"
 
 # 4x5 merra2 fullchem_aciduptake
-create_rundir "1\n5\n1\n1\n1\n${itRoot}\n\nn\n" "${log}" "${itRoot}"
+create_rundir "1\n5\n1\n1\n1\n${rundirsDir}\n\nn\n" "${log}"
 
 # 4x5 merra2 fullchem_APM
-create_rundir "1\n7\n1\n1\n1\n${itRoot}\n\nn\n" "${log}" "${itRoot}"
+create_rundir "1\n7\n1\n1\n1\n${rundirsDir}\n\nn\n" "${log}"
 
 # 4x5 merra2 fullchem_benchmark
-create_rundir "1\n2\n1\n1\n1\n${itRoot}\n\nn\n" "${log}" "${itRoot}"
+create_rundir "1\n2\n1\n1\n1\n${rundirsDir}\n\nn\n" "${log}"
 
 # 4x5 merra2 fullchem_complexSOA
-create_rundir "1\n3\n1\n1\n1\n1\n${itRoot}\n\nn\n" "${log}" "${itRoot}"
+create_rundir "1\n3\n1\n1\n1\n1\n${rundirsDir}\n\nn\n" "${log}"
 
 # 4x5 merra2 fullchem_complexSOA_SVPOA
-create_rundir "1\n3\n2\n1\n1\n1\n${itRoot}\n\nn\n" "${log}" "${itRoot}"
+create_rundir "1\n3\n2\n1\n1\n1\n${rundirsDir}\n\nn\n" "${log}"
 
 # 4x5 merra2 fullchem_marinePOA
-create_rundir "1\n4\n1\n1\n1\n${itRoot}\n\nn\n" "${log}" "${itRoot}"
+create_rundir "1\n4\n1\n1\n1\n${rundirsDir}\n\nn\n" "${log}"
 
 # 4x5 merra2 fullchem_RRTMG
-create_rundir "1\n8\n1\n1\n1\n${itRoot}\n\nn\n" "${log}" "${itRoot}"
+create_rundir "1\n8\n1\n1\n1\n${rundirsDir}\n\nn\n" "${log}"
 
 # 4x5 merra2 fullchem_TOMAS15_47L
-create_rundir "1\n6\n1\n1\n1\n2\n${itRoot}\n\nn\n" "${log}" "${itRoot}"
+create_rundir "1\n6\n1\n1\n1\n2\n${rundirsDir}\n\nn\n" "${log}"
 
 # 4x5 merra2 fullchem_TOMAS40_47L
-create_rundir "1\n6\n2\n1\n1\n2\n${itRoot}\n\nn\n" "${log}" "${itRoot}"
+create_rundir "1\n6\n2\n1\n1\n2\n${rundirsDir}\n\nn\n" "${log}"
 
 # 4x5 merra2 Hg
-create_rundir "5\n1\n1\n1\n${itRoot}\n\nn\n" "${log}" "${itRoot}"
+create_rundir "5\n1\n1\n1\n${rundirsDir}\n\nn\n" "${log}"
 
 # 4x5 merra2 POPs_BaP
-create_rundir "6\n1\n1\n1\n1\n${itRoot}\n\nn\n" "${log}" "${itRoot}"
+create_rundir "6\n1\n1\n1\n1\n${rundirsDir}\n\nn\n" "${log}"
 
 # 4x5 merra2 tagCH4
-create_rundir "7\n1\n1\n1\n${itRoot}\n\nn\n" "${log}" "${itRoot}"
+create_rundir "7\n1\n1\n1\n${rundirsDir}\n\nn\n" "${log}"
 
 # 4x5 merra2 tagCO
-create_rundir "8\n1\n1\n1\n${itRoot}\n\nn\n" "${log}" "${itRoot}"
+create_rundir "8\n1\n1\n1\n${rundirsDir}\n\nn\n" "${log}"
 
 # 4x5 merra2 tagO3
-create_rundir "9\n1\n1\n1\n${itRoot}\n\nn\n" "${log}" "${itRoot}"
+create_rundir "9\n1\n1\n1\n${rundirsDir}\n\nn\n" "${log}"
 
 # 4x5 merra2 TransportTracers
-create_rundir "10\n1\n1\n1\n${itRoot}\n\nn\n" "${log}" "${itRoot}"
+create_rundir "10\n1\n1\n1\n${rundirsDir}\n\nn\n" "${log}"
 
 # 4x5 merra2 TransportTracers_LuoWd
 dir="gc_4x5_merra2_TransportTracers_LuoWd"
-create_rundir "10\n1\n1\n1\n${itRoot}\n${dir}\nn\n" "${log}" "${itRoot}"
+create_rundir "10\n1\n1\n1\n${rundirsDir}\n${dir}\nn\n" "${log}"
 
 # 4x5 merra2 metals
-create_rundir "11\n1\n1\n1\n${itRoot}\n\nn\n" "${log}" "${itRoot}"
+create_rundir "11\n1\n1\n1\n${rundirsDir}\n\nn\n" "${log}"
 
 #=============================================================================
 # Create individual run directories: 4x5 and 47L (MERRA2)
 #=============================================================================
 
 # 4x5 merra2 fullchem_47L
-create_rundir "1\n1\n1\n1\n2\n${itRoot}\n\nn\n" "${log}" "${itRoot}"
+create_rundir "1\n1\n1\n1\n2\n${rundirsDir}\n\nn\n" "${log}"
 
 #=============================================================================
 # Nested-grid simulations
 #=============================================================================
 
 # 05x0625 merra2 CH4_47L_na
-create_rundir "3\n1\n3\n4\n2\n${itRoot}\n\nn\n" "${log}" "${itRoot}"
+create_rundir "3\n1\n3\n4\n2\n${rundirsDir}\n\nn\n" "${log}"
 
 # 05x0625 merra2 fullchem_47L_na
-create_rundir "1\n1\n1\n3\n4\n2\n${itRoot}\n\nn\n" "${log}" "${itRoot}"
+create_rundir "1\n1\n1\n3\n4\n2\n${rundirsDir}\n\nn\n" "${log}"
 
 #=============================================================================
 # Cleanup and quit
@@ -213,11 +251,17 @@ create_rundir "1\n1\n1\n3\n4\n2\n${itRoot}\n\nn\n" "${log}" "${itRoot}"
 cd "${thisDir}"
 
 # Free local variables
+unset binDir
+unset buildDir
 unset dir
+unset envDir
 unset geosChemDir
 unset itRoot
 unset log
+unset logsDir
 unset superProjectDir
+unset rundirsDirsDir
+unset scriptsDir
 unset thisDir
 
 # Free imported variables